[advanced] 최적화된 렌더링 적용
