+++ src/context/item/index.ts
@@ -1,0 +1,26 @@
+import { createContext, useContext } from "react";
+
+interface Item {
+  id: number;
+  name: string;
+  category: string;
+  price: number;
+}
+
+export interface ItemContextType {
+  items: Item[];
+  addItems: () => void;
+}
+
+export const ItemContext = createContext<ItemContextType | undefined>(
+  undefined,
+);
+
+export const useItemContext = () => {
+  const context = useContext(ItemContext);
+  if (context === undefined) {
+    throw new Error("useItemContext must be used within an ItemProvider");
+  }
+
+  return context;
+};
