+++ src/context/user/index.ts
@@ -1,0 +1,18 @@
+import { createContext, useContext } from "react";
+import { User } from "../../types/user";
+
+export interface UserType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export const UserContext = createContext<UserType | undefined>(undefined);
+
+export const useUserContext = () => {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error("useThemeContext must be used within an ThemeProvider");
+  }
+  return context;
+};
